# fr-administration-back/src/swager-specification/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All the users."
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getOneUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateOneUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteOneUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been successfully deleted."
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/roles": {
      "get": {
        "operationId": "UsersController_getRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/roles/users/{name}": {
      "get": {
        "operationId": "RolesController_usersByRoleName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/roles/{user}/{association}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "association",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "association",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "association",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Gestion des Associations",
    "description": "Descriptions des APIs de la gestion des associations",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UpdateUser": {
        "type": "object",
        "properties": {
          "lastname": {
            "type": "string",
            "description": "The last name of the user",
            "example": "Doe"
          },
          "firstname": {
            "type": "string",
            "description": "The first name of the user",
            "example": "John"
          },
          "age": {
            "type": "number",
            "description": "The age of the user",
            "example": 42
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "user@mail.com"
          }
        }
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "lastname": {
            "type": "string",
            "description": "The last name of the user",
            "example": "Doe"
          },
          "firstname": {
            "type": "string",
            "description": "The first name of the user",
            "example": "John"
          },
          "age": {
            "type": "number",
            "description": "The age of the user",
            "example": 42,
            "minimum": 0
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "user@mail.com"
          }
        },
        "required": [
          "lastname",
          "firstname",
          "age",
          "password",
          "email"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the roles of the given user in the given association",
            "example": "President"
          },
          "idUser": {
            "type": "number",
            "description": "The id of the user",
            "example": "1"
          },
          "idAssociation": {
            "type": "number",
            "description": "The id of the association",
            "example": "1"
          }
        },
        "required": [
          "name",
          "idUser",
          "idAssociation"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "lastname",
          "firstname",
          "age",
          "password"
        ]
      },
      "Association": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "required": [
          "id",
          "name",
          "users"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "number"
          },
          "associationId": {
            "type": "number"
          },
          "association": {
            "$ref": "#/components/schemas/Association"
          }
        },
        "required": [
          "name",
          "user",
          "userId",
          "associationId",
          "association"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The new name of the roles of the given user in the given association",
            "example": "President"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}